Fixed some Spec issues, fixed some marker issues, finished loadouts
