+++ CfgLoadouts.hpp
@@ -12,31 +12,31 @@ class CfgLoadouts {
     useFallback = 1;
 
     //Shared items
-    #include "Loadouts\common.hpp" // DO NOT COMMENT OUT
+    #include "Loadouts\common.hpp" // DO NOT COMMENT OUT, WILL BREAK EVERYTHING
 
     //Only include one hpp per faction; use (//) to comment out other files
 
     //BLUFOR FACTION (blu_f):
-    #include "Loadouts\bluf_us_m4_ocp.hpp"            // US: M4 - Tan
-    // #include "Loadouts\bluf_us_m4_ucp.hpp"         // US: M4 - Gray/Green
-    // #include "Loadouts\bluf_brit_l85_mtp.hpp"      // British: L86 - Multi-Terrain Pattern
-    // #include "Loadouts\bluf_swede_ak5_m90.hpp"     // Swedish: AK5 - Fish
-    // #include "Loadouts\bluf_ger_g36_fleck.hpp"     // German: G36 - Flecktarn Camo
+    #include "Loadouts\blu_us_m4_ucp.hpp"            // US: M4 - Gray/Green
+    // #include "Loadouts\blu_us_m4_ocp.hpp"         // US: M4 - Tan
+    // #include "Loadouts\blu_brit_l85_mtp.hpp"      // British: L86 - Multi-Terrain Pattern
+    // #include "Loadouts\blu_swede_ak5_m90.hpp"     // Swedish: AK5 - Fish
+    // #include "Loadouts\blu_ger_g36_fleck.hpp"     // German: G36 - Flecktarn Camo
 
     //INDFOR FACTION (ind_f):
-    // #include "Loadouts\indf_ukr_ak74_ttsko.hpp"       // Ukraine: AK74 - TTskO
-    // #include "Loadouts\indf_ukr_ak74_ddpm.hpp"     // "Ukraine": AK74 - Desert DPM
-    // #include "Loadouts\indf_reb_ak74_desert.hpp"   // Rebel: AK74 - Mixed Desert
-    #include "Loadouts\indf_ger_g36_tropen.hpp"    // German: G36 - Tropen Camo
+    #include "Loadouts\ind_ukr_ak74_ttsko.hpp"       // Ukraine: AK74 - TTskO
+    // #include "Loadouts\ind_ukr_ak74_ddpm.hpp"     // "Ukraine": AK74 - Desert DPM
+    // #include "Loadouts\ind_reb_ak47_desert.hpp"   // Rebel: AK47 - Mixed Desert
+    // #include "Loadouts\ind_ger_g36_tropen.hpp"    // German: G36 - Tropen Camo
 
     //OPFOR FACTION (opf_f):
-    // #include "Loadouts\opfor_desreb.hpp"         //Our standard "middle-east" loadout
-    // #include "Loadouts\opfor_standard.hpp"          //Our standard "middle-east" loadout
+    #include "Loadouts\opf_reb_ak47_desert.hpp"      // Rebel: AK47 - Mixed Desert
+    // #include "Loadouts\opf_ru_ak74_floral.hpp"    // Russian: AK74 - Floral
 
     //RUSSIAN FACTIONS (rhs_faction_msv):
-    // #include "Loadouts\russian_standard.hpp"     //Our standard "commy" loadout
-    // #include "Loadouts\russian_sov-sweedish.hpp"    //Soviet's "sweedish" loadout
+    // #include "Loadouts\msv_ru_ak74_emr.hpp"          //Our standard "commy" loadout
+    #include "Loadouts\msv_swede_ak5_m90k.hpp"    //Soviet's "sweedish" loadout
 
     //Civilians (mainly for RP missions)
-    #include "Loadouts\civilians.hpp"               //Bare example of doing civilian loadouts
+    #include "Loadouts\civilians.hpp"                //Bare example of doing civilian loadouts
 };
