+++ Loadouts/blu_ger_g36_fleck.hpp
@@ -1,65 +1,65 @@
 //Author: BWMF
-//Description: US OCP - M4 Rifles
+//Description: German Flekterarian Camo G36
 
 class blu_f {
   //Rifle
-  #define WEST_RIFLE "rhs_weap_m4a1_carryhandle_grip2"
-  #define WEST_RIFLE_MAG "rhs_mag_30Rnd_556x45_Mk318_Stanag:8","rhs_mag_30Rnd_556x45_M855A1_Stanag_Tracer_Red:2"
-  #define WEST_RIFLE_ATTACHMENTS "rhsusf_acc_eotech_552","rhsusf_acc_anpeq15A"
+  #define WEST_RIFLE "BWA3_G36"
+  #define WEST_RIFLE_MAG "BWA3_30Rnd_556x45_G36:8","BWA3_30Rnd_556x45_G36_Tracer:2"
+  #define WEST_RIFLE_ATTACHMENTS "BWA3_optic_EOTech","BWA3_acc_LLM01_irlaser"
   #define WEST_AAR_ATTACHMENTS WEST_RIFLE_ATTACHMENTS
   //GL Rifle
-  #define WEST_GLRIFLE "rhs_weap_m4a1_carryhandle_m203S"
+  #define WEST_GLRIFLE "BWA3_G36_AG"
   #define WEST_GLRIFLE_MAG WEST_RIFLE_MAG
   #define WEST_GLRIFLE_MAG_SMOKE "1Rnd_Smoke_Grenade_shell:2","1Rnd_SmokeRed_Grenade_shell:2"
-  #define WEST_GLRIFLE_MAG_HE "rhs_mag_M433_HEDP:5"
+  #define WEST_GLRIFLE_MAG_HE "1Rnd_HE_Grenade_shell:5"
   #define WEST_GLRIFLE_MAG_FLARE "UGL_FlareYellow_F:4"
   //Carbine
-  #define WEST_CARBINE "rhs_weap_m4a1_carryhandle_grip"
+  #define WEST_CARBINE "BWA3_G36K"
   #define WEST_CARBINE_MAG WEST_RIFLE_MAG
   // AR
-  #define WEST_AR "rhs_weap_m249_pip_L"
-  #define WEST_AR_MAG "rhsusf_200Rnd_556x45_soft_pouch:2"
+  #define WEST_AR "BWA3_MG4"
+  #define WEST_AR_MAG "BWA3_200Rnd_556x45:2"
   // AT
-  #define WEST_AT "rhs_weap_M136"
+  #define WEST_AT "BWA3_Pzf3"
+  #define WEST_AT_MAG "BWA3_Pzf3_IT"
   // MMG
-  #define WEST_MMG "rhs_weap_m240B"
-  #define WEST_MMG_MAG "rhsusf_100Rnd_762x51:5"
-  #define WEST_MMG_MAG2 "rhsusf_100Rnd_762x51:4"
+  #define WEST_MMG "BWA3_MG5"
+  #define WEST_MMG_MAG "BWA3_120Rnd_762x51:4"
   // Sniper Rifle
-  #define WEST_SNIPER "rhs_weap_m14ebrri"
-  #define WEST_SNIPER_MAG "rhsusf_20Rnd_762x51_m993_Mag:8"
+  #define WEST_SNIPER "BWA3_G28_Standard"
+  #define WEST_SNIPER_MAG "BWA3_10Rnd_762x51_G28:10"
   // Spotter Rifle
-  #define WEST_SPOTTER "rhs_weap_m4a1_blockII_grip2_KAC"
-  #define WEST_SPOTTER_MAG "rhs_mag_30Rnd_556x45_Mk318_Stanag:8"
+  #define WEST_SPOTTER "BWA3_G36"
+  #define WEST_SPOTTER_MAG "BWA3_30Rnd_556x45_G36:8"
   // SMG
-  #define WEST_SMG "hlc_smg_mp5a4"
-  #define WEST_SMG_MAG "hlc_30Rnd_9x19_B_MP5:6"
+  #define IND_SMG "BWA3_MP7"
+  #define IND_SMG_MAG "BWA3_40Rnd_46x30_MP7:6"
   // Pistol
-  #define WEST_PISTOL "rhsusf_weap_m1911a1"
-  #define WEST_PISTOL_MAG "rhsusf_mag_7x45acp_MHP:3"
+  #define IND_PISTOL "BWA3_P8"
+  #define IND_PISTOL_MAG "BWA3_15Rnd_9x19_P8:3"
 
   class Car {
     TransportWeapons[] = {WEST_AT};
-    TransportMagazines[] = {WEST_RIFLE_MAG,WEST_RIFLE_MAG,WEST_CARBINE_MAG,WEST_AR_MAG,WEST_AR_MAG,WEST_GLRIFLE_MAG_HE};
-    TransportItems[] = {BASE_MEDICAL,BASE_MEDICAL,BASE_MEDICAL,BASE_MEDICAL};
+    TransportMagazines[] = {WEST_RIFLE_MAG,WEST_RIFLE_MAG,WEST_CARBINE_MAG,WEST_AR_MAG,WEST_AR_MAG,WEST_GLRIFLE_MAG_HE,WEST_AT_MAG};
+    TransportItems[] = {WEST_BASE_MEDICAL,WEST_BASE_MEDICAL,WEST_BASE_MEDICAL,WEST_BASE_MEDICAL};
   };
   class Tank {
     TransportWeapons[] = {WEST_AT};
-    TransportMagazines[] = {WEST_RIFLE_MAG,WEST_RIFLE_MAG,WEST_CARBINE_MAG,WEST_AR_MAG,WEST_AR_MAG,WEST_GLRIFLE_MAG_HE};
-    TransportItems[] = {BASE_MEDICAL,BASE_MEDICAL,BASE_MEDICAL,BASE_MEDICAL};
+    TransportMagazines[] = {WEST_RIFLE_MAG,WEST_RIFLE_MAG,WEST_CARBINE_MAG,WEST_AR_MAG,WEST_AR_MAG,WEST_GLRIFLE_MAG_HE,WEST_AT_MAG};
+    TransportItems[] = {WEST_BASE_MEDICAL,WEST_BASE_MEDICAL,WEST_BASE_MEDICAL,WEST_BASE_MEDICAL};
   };
   class Helicopter {
     TransportMagazines[] = {WEST_RIFLE_MAG,WEST_RIFLE_MAG,WEST_CARBINE_MAG,WEST_AR_MAG,WEST_AR_MAG,WEST_GLRIFLE_MAG_HE};
-    TransportItems[] = {BASE_MEDICAL,BASE_MEDICAL,BASE_MEDICAL,BASE_MEDICAL};
+    TransportItems[] = {WEST_BASE_MEDICAL,WEST_BASE_MEDICAL,WEST_BASE_MEDICAL,WEST_BASE_MEDICAL};
   };
   class Plane {};
   class Ship_F {};
 
   class B_Soldier_F {// rifleman
-    uniform[] = {"rhs_uniform_cu_ocp"};
-    vest[] = {"rhsusf_iotv_ocp_Rifleman"};
-    headgear[] = {"rhsusf_ach_helmet_ocp","rhsusf_ach_helmet_ESS_ocp"};
-    backpack[] = {"rhsusf_assault_eagleaiii_ocp"};
+    uniform[] = {"BWA3_Uniform_Fleck","BWA3_Uniform2_Fleck"};
+    vest[] = {"BWA3_Vest_Fleck","BWA3_Vest_Rifleman1_Fleck"};
+    headgear[] = {"BWA3_OpsCore_Fleck","BWA3_OpsCore_Fleck_Patch","BWA3_OpsCore_Fleck_Camera","BWA3_OpsCore_Schwarz"};
+    backpack[] = {"BWA3_Kitbag_Fleck"};
     backpackItems[] = {BASE_MEDICAL};
     weapons[] = {WEST_RIFLE};
     magazines[] = {WEST_RIFLE_MAG,WEST_BASE_GRENADES};
@@ -72,15 +72,13 @@ class blu_f {
     magazines[] = {WEST_CARBINE_MAG,WEST_BASE_GRENADES};
   };
   class B_Soldier_TL_F: B_Soldier_F {// FTL
-    vest[] = {"rhsusf_iotv_ocp_Grenadier"};
+    vest[] = {"BWA3_Vest_Leader_Fleck","BWA3_Vest_Grenadier_Fleck"};
     weapons[] = {WEST_GLRIFLE};
-    headgear[] = {"rhsusf_ach_helmet_headset_ess_ocp"};
     magazines[] = {WEST_GLRIFLE_MAG,WEST_GLRIFLE_MAG_HE,WEST_GLRIFLE_MAG_SMOKE,WEST_LEADER_GRENADES};
     items[] += {WEST_LEADER_TOOLS};
     linkedItems[] += {WEST_LEADER_LINKED,BINOS};
   };
   class B_Soldier_SL_F: B_Soldier_TL_F {// SL
-    headgear[] = {"rhsusf_ach_helmet_headset_ocp"};
     handguns[] = {WEST_PISTOL};
     magazines[] += {WEST_PISTOL_MAG};
     linkedItems[] = {WEST_BASE_LINKED,WEST_LEADER_LINKED,RANGE_FINDER};
@@ -94,7 +92,7 @@ class blu_f {
     linkedItems[] += {"B_uavterminal"};
   };
   class B_Soldier_AR_F: B_Soldier_F {// AR
-    vest[] = {"rhsusf_iotv_ocp_SAW"};
+    vest[] = {"BWA3_Vest_Autorifleman_Fleck"};
     weapons[] = {WEST_AR};
     magazines[] = {WEST_AR_MAG,WEST_PISTOL_MAG,WEST_BASE_GRENADES};
     handguns[] = {WEST_PISTOL};
@@ -105,10 +103,12 @@ class blu_f {
     linkedItems[] += {BINOS};
   };
   class B_Soldier_LAT_F: Fic_Soldier_Carbine {// RAT
+    magazines[] += {WEST_AT_MAG};
     launchers[] = {WEST_AT};
   };
   class B_medic_F: Fic_Soldier_Carbine {// Medic
-    vest[] = {"rhsusf_iotv_ocp_medic"};
+    vest[] = {"BWA3_Vest_Medic_Fleck"};
+    backpack[] = {"BWA3_Kitbag_Fleck_Medic"};
     magazines[] = {WEST_CARBINE_MAG,MEDIC_GRENADES};
     backpackItems[] = {MEDIC_MEDICAL};
   };
@@ -121,22 +121,22 @@ class blu_f {
     attachments[] = {};
   };
   class B_Soldier_A_F: Fic_Spotter {// MMG Spotter/Ammo Bearer
-    backpackItems[] += {WEST_MMG_MAG2};
+    backpackItems[] += {WEST_MMG_MAG};
   };
   class B_soldier_AT_F: Fic_Soldier_Carbine {// MAT Gunner
-    LAUNCHER_GEAR("B_Carryall_mcamo", WEST_MAT_MAG)
+    LAUNCHER_GEAR("BWA3_Carryall_Fleck", WEST_MAT_MAG)
     launchers[] = {WEST_MAT};
     secondaryAttachments[] = {WEST_MAT_OPTIC};
   };
   class B_Soldier_AAT_F: Fic_Spotter {// MAT Spotter/Ammo Bearer
-    LAUNCHER_GEAR("B_Carryall_mcamo", WEST_MAT_MAG)
+    LAUNCHER_GEAR("BWA3_Carryall_Fleck", WEST_MAT_MAG)
   };
   class B_soldier_AA_F: Fic_Soldier_Carbine {// SAM Gunner
-    LAUNCHER_GEAR("B_Carryall_mcamo", WEST_SAM_MAG)
+    LAUNCHER_GEAR("BWA3_Carryall_Fleck", WEST_SAM_MAG)
     launchers[] = {WEST_SAM};
   };
   class B_Soldier_AAA_F: Fic_Spotter {// SAM Spotter/Ammo Bearer
-    LAUNCHER_GEAR("B_Carryall_mcamo", WEST_SAM_MAG2)
+    LAUNCHER_GEAR("BWA3_Carryall_Fleck", WEST_SAM_MAG2)
   };
   class B_support_Mort_F: Fic_Soldier_Carbine {// Mortar Gunner
     MORTAR_GEAR("B_Mortar_01_weapon_F")
@@ -145,8 +145,9 @@ class blu_f {
     MORTAR_GEAR("B_Mortar_01_support_F")
   };
   class B_spotter_F: Fic_Spotter {// Spotter
-    vest[] = {"rhsusf_iotv_ocp"};
-    headgear[] = {"rhs_Booniehat_ocp"};
+    uniform[] = {"BWA3_Uniform_Ghillie_Fleck"};
+    vest[] = {"BWA3_Vest_Marksman_Fleck"};
+    headgear[] = {};
     weapons[] = {WEST_SPOTTER};
     magazines[] = {WEST_SPOTTER_MAG,WEST_BASE_GRENADES};
     items[] += {RADIO_MR,"ACE_ATragMX","ACE_Kestrel4500","ACE_RangeCard"};
@@ -157,7 +158,7 @@ class blu_f {
     magazines[] = {WEST_SNIPER_MAG,WEST_BASE_GRENADES};
     items[] = {BASE_MEDICAL,BASE_TOOLS,"ACE_RangeCard"};
     linkedItems[] = {WEST_BASE_LINKED};
-    attachments[] = {"rhsusf_acc_LEUPOLDMK4_2","rhsusf_acc_anpeq15A"};
+    attachments[] = {"BWA3_optic_Shortdot","BWA3_acc_VarioRay_irlaser"};
   };
   class B_Helipilot_F {// Pilot
     uniform[] = {"U_B_HeliPilotCoveralls"};
@@ -172,23 +173,22 @@ class blu_f {
   };
   class B_helicrew_F: B_Helipilot_F {}; // Pilot
   class B_crew_F: Fic_Soldier_Carbine {// Crew
-    vest[] = {"rhsusf_iotv_ocp"};
-    headgear[] = {"rhsusf_cvc_green_ess"};
-    backpack[] = {"B_Carryall_mcamo"};
+    uniform[] = {"BWA3_Uniform_Crew_Fleck"};
+    headgear[] = {"BWA3_CrewmanKSK_Fleck_Headset"};
+    backpack[] = {"BWA3_Kitbag_Fleck"};
     magazines[] = {WEST_CARBINE_MAG,CREW_GRENADES};
     backpackItems[] = {KEY_WEST,RADIO_LR};
     linkedItems[] = {WEST_BASE_LINKED,WEST_LEADER_LINKED,BINOS};
+    items[] += {BASE_MEDICAL};
   };
   class B_soldier_repair_F: B_crew_F {// Repair Specialist
-    backpack[] = {"B_Carryall_mcamo"};
     backpackItems[] = {"Toolkit",RADIO_MR,KEY_WEST};
-    vest[] = {"rhsusf_iotv_ocp_repair"};
     linkedItems[] = {WEST_BASE_LINKED,WEST_LEADER_LINKED};
   };
   class Fic_eng: B_soldier_repair_F {
-    headgear[] = {"rhsusf_ach_helmet_ocp","rhsusf_ach_helmet_ESS_ocp"};
-    items[] += {BASE_ENG};
+    backpack[] = {"BWA3_PatrolPack_Fleck"};
     backpackItems[] = {};
+    items[] += {BASE_ENG};
   };
   class B_soldier_exp_F: Fic_eng {// Explosive Specialist
     magazines[] += {BASE_EXP};
