+++ Loadouts/blu_swede_ak5_m90.hpp
@@ -1,42 +1,42 @@
 //Author: BWMF
-//Description: Red Coats
+//Description: Swedish Feesh
 
 class blu_f {
   //Rifle
-  #define WEST_RIFLE "UK3CB_BAF_L85A2_RIS","UK3CB_BAF_L85A2_RIS_AFG"
-  #define WEST_RIFLE_MAG "UK3CB_BAF_30Rnd:8","UK3CB_BAF_30Rnd_T:2"
-  #define WEST_RIFLE_ATTACHMENTS "UK3CB_BAF_Eotech","UK3CB_BAF_LLM_IR_Black"
+  #define WEST_RIFLE "SAM_AK5C","SAM_AK5C_NOGRIP"
+  #define WEST_RIFLE_MAG "SAM_30Rnd_556x45_Pmag:8","SAM_30Rnd_556x45_Pmag_T:2"
+  #define WEST_RIFLE_ATTACHMENTS "rhsusf_acc_eotech_552","rhsusf_acc_anpeq15A"
   #define WEST_AAR_ATTACHMENTS WEST_RIFLE_ATTACHMENTS
   //GL Rifle
-  #define WEST_GLRIFLE "UK3CB_BAF_L85A2_UGL_HWS","UK3CB_BAF_L85A2_UGL"
+  #define WEST_GLRIFLE "SAM_AK5C_GRT"
   #define WEST_GLRIFLE_MAG WEST_RIFLE_MAG
   #define WEST_GLRIFLE_MAG_SMOKE "1Rnd_Smoke_Grenade_shell:2","1Rnd_SmokeRed_Grenade_shell:2"
   #define WEST_GLRIFLE_MAG_HE "1Rnd_HE_Grenade_shell:5"
   #define WEST_GLRIFLE_MAG_FLARE "UGL_FlareYellow_F:4"
   //Carbine
-  #define WEST_CARBINE WEST_RIFLE
+  #define WEST_CARBINE "SAM_AK5C","SAM_AK5C_NOGRIP"
   #define WEST_CARBINE_MAG WEST_RIFLE_MAG
   // AR
-  #define WEST_AR "UK3CB_BAF_L110A2"
-  #define WEST_AR_MAG "UK3CB_BAF_100Rnd:4"
+  #define WEST_AR "SAM_KSP90C"
+  #define WEST_AR_MAG "SAM_200Rnd_556x45:2"
   // AT
-  #define WEST_AT "UK3CB_BAF_AT4_AT_Launcher"
+  #define WEST_AT "rhs_weap_M136"
+  #define WEST_AT_MAG "rhs_m136_mag"
   // MMG
-  #define WEST_MMG "UK3CB_BAF_L7A2"
-  #define WEST_MMG_MAG "UK3CB_BAF_75Rnd:6"
-  #define WEST_MMG_MAG2 "UK3CB_BAF_75Rnd:5"
+  #define WEST_MMG "SAM_KSP94"
+  #define WEST_MMG_MAG "SAM_120Rnd_762x51:3","SAM_120Rnd_762x51_T:1"
   // Sniper Rifle
-  #define WEST_SNIPER "UK3CB_BAF_L82A1"
-  #define WEST_SNIPER_MAG "UK3CB_BAF_10Rnd_127x99mm:4"
+  #define WEST_SNIPER "rhs_weap_m14ebrri"
+  #define WEST_SNIPER_MAG "rhsusf_20Rnd_762x51_m993_Mag:8"
   // Spotter Rifle
-  #define WEST_SPOTTER "UK3CB_BAF_L86A3"
-  #define WEST_SPOTTER_MAG "UK3CB_BAF_30Rnd:8"
+  #define WEST_SPOTTER "rhs_weap_m4a1_blockII_grip2_KAC"
+  #define WEST_SPOTTER_MAG "rhs_mag_30Rnd_556x45_Mk318_Stanag:8"
   // SMG
-  #define WEST_SMG "UK3CB_BAF_L91A1"
-  #define WEST_SMG_MAG "UK3CB_BAF_30Rnd_9mm:6"
+  #define WEST_SMG "hlc_smg_mp5a4"
+  #define WEST_SMG_MAG "hlc_30Rnd_9x19_B_MP5:6"
   // Pistol
-  #define WEST_PISTOL "UK3CB_BAF_L131A1"
-  #define WEST_PISTOL_MAG "UK3CB_BAF_17Rnd_9mm:3"
+  #define WEST_PISTOL "rhsusf_weap_m1911a1"
+  #define WEST_PISTOL_MAG "rhsusf_mag_7x45acp_MHP:3"
 
   class Car {
     TransportWeapons[] = {WEST_AT};
@@ -46,20 +46,20 @@ class blu_f {
   class Tank {
     TransportWeapons[] = {WEST_AT};
     TransportMagazines[] = {WEST_RIFLE_MAG,WEST_RIFLE_MAG,WEST_CARBINE_MAG,WEST_AR_MAG,WEST_AR_MAG,WEST_GLRIFLE_MAG_HE};
-    TransportItems[] = {WEST_BASE_MEDICAL,WEST_BASE_MEDICAL,WEST_BASE_MEDICAL,WEST_BASE_MEDICAL};
+    TransportItems[] = {BASE_MEDICAL,BASE_MEDICAL,BASE_MEDICAL,BASE_MEDICAL};
   };
   class Helicopter {
     TransportMagazines[] = {WEST_RIFLE_MAG,WEST_RIFLE_MAG,WEST_CARBINE_MAG,WEST_AR_MAG,WEST_AR_MAG,WEST_GLRIFLE_MAG_HE};
-    TransportItems[] = {WEST_BASE_MEDICAL,WEST_BASE_MEDICAL,WEST_BASE_MEDICAL,WEST_BASE_MEDICAL};
+    TransportItems[] = {BASE_MEDICAL,BASE_MEDICAL,BASE_MEDICAL,BASE_MEDICAL};
   };
   class Plane {};
   class Ship_F {};
 
   class B_Soldier_F {// rifleman
-    uniform[] = {"UK3CB_BAF_U_CombatUniform_MTP","UK3CB_BAF_U_CombatUniform_MTP_ShortSleeve"};
-    vest[] = {"UK3CB_BAF_V_Osprey_Rifleman_A","UK3CB_BAF_V_Osprey_Rifleman_B","UK3CB_BAF_V_Osprey_Rifleman_C","UK3CB_BAF_V_Osprey_Rifleman_D","UK3CB_BAF_V_Osprey_Rifleman_E","UK3CB_BAF_V_Osprey_Rifleman_F"};
-    headgear[] = {"UK3CB_BAF_H_Mk7_Camo_A","UK3CB_BAF_H_Mk7_Camo_B","UK3CB_BAF_H_Mk7_Camo_C","UK3CB_BAF_H_Mk7_Camo_D","UK3CB_BAF_H_Mk7_Camo_E","UK3CB_BAF_H_Mk7_Camo_F","UK3CB_BAF_H_Mk7_Camo_ESS_A","UK3CB_BAF_H_Mk7_Camo_ESS_B","UK3CB_BAF_H_Mk7_Camo_ESS_C","UK3CB_BAF_H_Mk7_Camo_ESS_D","UK3CB_BAF_H_Mk7_Camo_CESS_A","UK3CB_BAF_H_Mk7_Camo_CESS_B","UK3CB_BAF_H_Mk7_Camo_CESS_C"};
-    backpack[] = {"UK3CB_BAF_B_Bergen_MTP_Rifleman_L_A","UK3CB_BAF_B_Bergen_MTP_Rifleman_L_B","UK3CB_BAF_B_Bergen_MTP_Rifleman_L_C","UK3CB_BAF_B_Bergen_MTP_Rifleman_L_D"};
+    uniform[] = {"SAM_Crye_M90_uniform","SAM_Crye_M90_uniform_r"};
+    vest[] = {"SAM_PlateCarrier"};
+    headgear[] = {"SAM_HelmetIA_M90"};
+    backpack[] = {"SAM_AssaultPack_rgr"};
     backpackItems[] = {BASE_MEDICAL};
     weapons[] = {WEST_RIFLE};
     magazines[] = {WEST_RIFLE_MAG,WEST_BASE_GRENADES};
@@ -72,15 +72,12 @@ class blu_f {
     magazines[] = {WEST_CARBINE_MAG,WEST_BASE_GRENADES};
   };
   class B_Soldier_TL_F: B_Soldier_F {// FTL
-    backpack[] = {"UK3CB_BAF_B_Bergen_MTP_Radio_L_B","UK3CB_BAF_B_Bergen_MTP_Radio_L_A","UK3CB_BAF_B_Bergen_MTP_PointMan_L_A"};
     weapons[] = {WEST_GLRIFLE};
     magazines[] = {WEST_GLRIFLE_MAG,WEST_GLRIFLE_MAG_HE,WEST_GLRIFLE_MAG_SMOKE,WEST_LEADER_GRENADES};
     items[] += {WEST_LEADER_TOOLS};
     linkedItems[] += {WEST_LEADER_LINKED,BINOS};
   };
   class B_Soldier_SL_F: B_Soldier_TL_F {// SL
-    vest[] = {"UK3CB_BAF_V_Osprey_SL_D","UK3CB_BAF_V_Osprey_SL_C","UK3CB_BAF_V_Osprey_SL_B","UK3CB_BAF_V_Osprey_SL_A"};
-    backpack[] = {"UK3CB_BAF_B_Bergen_MTP_SL_L_A","UK3CB_BAF_B_Bergen_MTP_SL_H_A","UK3CB_BAF_B_Bergen_MTP_JTAC_L_A"};
     handguns[] = {WEST_PISTOL};
     magazines[] += {WEST_PISTOL_MAG};
     linkedItems[] = {WEST_BASE_LINKED,WEST_LEADER_LINKED,RANGE_FINDER};
@@ -99,17 +96,16 @@ class blu_f {
     handguns[] = {WEST_PISTOL};
   };
   class B_Soldier_AAR_F: B_Soldier_F {// AAR
-    vest[] = {"UK3CB_BAF_V_Osprey_MG_A","UK3CB_BAF_V_Osprey_MG_B"};
     backpackItems[] += {WEST_AR_MAG};
     attachments[] = {WEST_AAR_ATTACHMENTS};
     linkedItems[] += {BINOS};
   };
   class B_Soldier_LAT_F: Fic_Soldier_Carbine {// RAT
+    magazines[] += {WEST_AT_MAG};
     launchers[] = {WEST_AT};
   };
   class B_medic_F: Fic_Soldier_Carbine {// Medic
-    vest[] = {"UK3CB_BAF_V_Osprey_Medic_A","UK3CB_BAF_V_Osprey_Medic_B","UK3CB_BAF_V_Osprey_Medic_C","UK3CB_BAF_V_Osprey_Medic_D"};
-    backpack[] = {"UK3CB_BAF_B_Bergen_MTP_Medic_L_A","UK3CB_BAF_B_Bergen_MTP_Medic_L_B"};
+    backpack[] = {"SAM_Kitbag_rgr_med"};
     magazines[] = {WEST_CARBINE_MAG,MEDIC_GRENADES};
     backpackItems[] = {MEDIC_MEDICAL};
   };
@@ -117,29 +113,29 @@ class blu_f {
     linkedItems[] += {RANGE_FINDER};
   };
   class B_support_MG_F: B_Soldier_AR_F {// MMG
-    backpack[] = {"UK3CB_BAF_B_Bergen_MTP_Rifleman_H_A"};
+    backpack[] = {"SAM_Kitbag_rgr"};
     weapons[] = {WEST_MMG};
     magazines[] = {WEST_MMG_MAG,WEST_PISTOL_MAG,WEST_BASE_GRENADES};
     attachments[] = {};
   };
   class B_Soldier_A_F: Fic_Spotter {// MMG Spotter/Ammo Bearer
-    backpack[] = {"UK3CB_BAF_B_Bergen_MTP_Rifleman_H_A"};
-    backpackItems[] += {WEST_MMG_MAG2};
+    backpack[] = {"SAM_Kitbag_rgr"};
+    backpackItems[] += {WEST_MMG_MAG};
   };
   class B_soldier_AT_F: Fic_Soldier_Carbine {// MAT Gunner
-    LAUNCHER_GEAR("UK3CB_BAF_B_Bergen_MTP_Rifleman_H_B", WEST_MAT_MAG)
+    LAUNCHER_GEAR("SAM_GRG_Backpack_Heat", WEST_MAT_MAG)
     launchers[] = {WEST_MAT};
     secondaryAttachments[] = {WEST_MAT_OPTIC};
   };
   class B_Soldier_AAT_F: Fic_Spotter {// MAT Spotter/Ammo Bearer
-    LAUNCHER_GEAR("UK3CB_BAF_B_Bergen_MTP_Rifleman_H_B", WEST_MAT_MAG)
+    LAUNCHER_GEAR("SAM_GRG_Backpack_Heat", WEST_MAT_MAG2)
   };
   class B_soldier_AA_F: Fic_Soldier_Carbine {// SAM Gunner
-    LAUNCHER_GEAR("UK3CB_BAF_B_Bergen_MTP_Rifleman_H_C", WEST_SAM_MAG)
+    LAUNCHER_GEAR("B_Carryall_oli", WEST_SAM_MAG)
     launchers[] = {WEST_SAM};
   };
   class B_Soldier_AAA_F: Fic_Spotter {// SAM Spotter/Ammo Bearer
-    LAUNCHER_GEAR("UK3CB_BAF_B_Bergen_MTP_Rifleman_H_C", WEST_SAM_MAG2)
+    LAUNCHER_GEAR("B_Carryall_oli", WEST_SAM_MAG2)
   };
   class B_support_Mort_F: Fic_Soldier_Carbine {// Mortar Gunner
     MORTAR_GEAR("B_Mortar_01_weapon_F")
@@ -148,10 +144,8 @@ class blu_f {
     MORTAR_GEAR("B_Mortar_01_support_F")
   };
   class B_spotter_F: Fic_Spotter {// Spotter
-    uniform[] = {"UK3CB_BAF_U_CombatUniform_MTP_Ghillie_RM"};
-    vest[] = {"UK3CB_BAF_V_Osprey_Marksman_A"};
-    headgear[] = {"UK3CB_BAF_H_Mk7_Scrim_ESS_A","UK3CB_BAF_H_Mk7_Scrim_ESS_B","UK3CB_BAF_H_Mk7_Scrim_ESS_C"};
-    backpack[] = {"UK3CB_BAF_B_Bergen_MTP_Rifleman_H_C"};
+    uniform[] = {"SAM_Crye_M90_ghillie"};
+    headgear[] = {"SAM_Watchcap_grn"};
     weapons[] = {WEST_SPOTTER};
     magazines[] = {WEST_SPOTTER_MAG,WEST_BASE_GRENADES};
     items[] += {RADIO_MR,"ACE_ATragMX","ACE_Kestrel4500","ACE_RangeCard"};
@@ -160,15 +154,15 @@ class blu_f {
   class B_sniper_F: B_spotter_F {// Sniper
     weapons[] = {WEST_SNIPER};
     magazines[] = {WEST_SNIPER_MAG,WEST_BASE_GRENADES};
-    items[] = {BASE_MEDICAL,BASE_TOOLS,"ACE_RangeCard"};
+    items[] = {BASE_MEDICAL,WEST_BASE_TOOLS,"ACE_RangeCard"};
     linkedItems[] = {WEST_BASE_LINKED};
-    attachments[] = {"UK3CB_BAF_SB31250_Ghillie"};
+    attachments[] = {"rhsusf_acc_LEUPOLDMK4_2","rhsusf_acc_anpeq15A"};
   };
   class B_Helipilot_F {// Pilot
-    uniform[] = {"UK3CB_BAF_U_HeliPilotCoveralls_RAF"};
+    uniform[] = {"U_B_HeliPilotCoveralls"};
     backpack[] = {"B_AssaultPack_blk"};
-    vest[] = {"UK3CB_BAF_V_Pilot_A"};
-    headgear[] = {"UK3CB_BAF_H_PilotHelmetHeli_A"};
+    vest[] = {"V_TacVest_blk"};
+    headgear[] = {"H_PilotHelmetHeli_B"};
     weapons[] = {WEST_SMG};
     magazines[] = {WEST_SMG_MAG,CREW_GRENADES};
     backpackItems[] = {KEY_WEST,RADIO_LR};
@@ -177,16 +171,14 @@ class blu_f {
   };
   class B_helicrew_F: B_Helipilot_F {}; // Pilot
   class B_crew_F: Fic_Soldier_Carbine {// Crew
-    uniform[] = {"UK3CB_BAF_U_CrewmanCoveralls_RTR"};
-    vest[] = {"UK3CB_BAF_V_Osprey","UK3CB_BAF_V_Osprey_Belt_A"};
-    headgear[] = {"UK3CB_BAF_H_CrewHelmet_A","UK3CB_BAF_H_CrewHelmet_ESS_A"};
-    backpack[] = {"UK3CB_BAF_B_Bergen_MTP_Engineer_L_A"};
+    backpack[] = {"SAM_Kitbag_rgr_eng"};
     magazines[] = {WEST_CARBINE_MAG,CREW_GRENADES};
     backpackItems[] = {KEY_WEST,RADIO_LR};
     linkedItems[] = {WEST_BASE_LINKED,WEST_LEADER_LINKED,BINOS};
+    items[] += {BASE_MEDICAL};
   };
   class B_soldier_repair_F: B_crew_F {// Repair Specialist
-    backpack[] = {"UK3CB_BAF_B_Bergen_MTP_Engineer_H_A"};
+    backpack[] = {"SAM_Kitbag_rgr_eng"};
     backpackItems[] = {"Toolkit",RADIO_MR,KEY_WEST};
     linkedItems[] = {WEST_BASE_LINKED,WEST_LEADER_LINKED};
   };
