+++ Loadouts/blu_us_m4_ucp.hpp
@@ -1,30 +1,31 @@
 //Author: BWMF
-//Description: Swedish Feesh
+//Description: US UCP - M4 Rifles
 
 class blu_f {
   //Rifle
-  #define WEST_RIFLE "SAM_AK5C","SAM_AK5C_NOGRIP"
-  #define WEST_RIFLE_MAG "SAM_30Rnd_556x45_Pmag:8","SAM_30Rnd_556x45_Pmag_T:2"
+  #define WEST_RIFLE "rhs_weap_m4a1_carryhandle_grip2"
+  #define WEST_RIFLE_MAG "rhs_mag_30Rnd_556x45_Mk318_Stanag:8","rhs_mag_30Rnd_556x45_M855A1_Stanag_Tracer_Red:2"
   #define WEST_RIFLE_ATTACHMENTS "rhsusf_acc_eotech_552","rhsusf_acc_anpeq15A"
   #define WEST_AAR_ATTACHMENTS WEST_RIFLE_ATTACHMENTS
   //GL Rifle
-  #define WEST_GLRIFLE "SAM_AK5C_GRT"
+  #define WEST_GLRIFLE "rhs_weap_m4a1_carryhandle_m203S"
   #define WEST_GLRIFLE_MAG WEST_RIFLE_MAG
   #define WEST_GLRIFLE_MAG_SMOKE "1Rnd_Smoke_Grenade_shell:2","1Rnd_SmokeRed_Grenade_shell:2"
-  #define WEST_GLRIFLE_MAG_HE "1Rnd_HE_Grenade_shell:5"
+  #define WEST_GLRIFLE_MAG_HE "rhs_mag_M433_HEDP:5"
   #define WEST_GLRIFLE_MAG_FLARE "UGL_FlareYellow_F:4"
   //Carbine
-  #define WEST_CARBINE "SAM_AK5C","SAM_AK5C_NOGRIP"
+  #define WEST_CARBINE "rhs_weap_m4a1_carryhandle_grip"
   #define WEST_CARBINE_MAG WEST_RIFLE_MAG
   // AR
-  #define WEST_AR "SAM_KSP90C"
-  #define WEST_AR_MAG "SAM_200Rnd_556x45:2"
+  #define WEST_AR "rhs_weap_m249_pip_L"
+  #define WEST_AR_MAG "rhsusf_200Rnd_556x45_soft_pouch:2"
   // AT
   #define WEST_AT "rhs_weap_M136"
   #define WEST_AT_MAG "rhs_m136_mag"
   // MMG
-  #define WEST_MMG "SAM_KSP94"
-  #define WEST_MMG_MAG "SAM_120Rnd_762x51:3","SAM_120Rnd_762x51_T:1"
+  #define WEST_MMG "rhs_weap_m240B"
+  #define WEST_MMG_MAG "rhsusf_100Rnd_762x51:5"
+  #define WEST_MMG_MAG2 "rhsusf_100Rnd_762x51:4"
   // Sniper Rifle
   #define WEST_SNIPER "rhs_weap_m14ebrri"
   #define WEST_SNIPER_MAG "rhsusf_20Rnd_762x51_m993_Mag:8"
@@ -41,7 +42,7 @@ class blu_f {
   class Car {
     TransportWeapons[] = {WEST_AT};
     TransportMagazines[] = {WEST_RIFLE_MAG,WEST_RIFLE_MAG,WEST_CARBINE_MAG,WEST_AR_MAG,WEST_AR_MAG,WEST_GLRIFLE_MAG_HE};
-    TransportItems[] = {WEST_BASE_MEDICAL,WEST_BASE_MEDICAL,WEST_BASE_MEDICAL,WEST_BASE_MEDICAL};
+    TransportItems[] = {BASE_MEDICAL,BASE_MEDICAL,BASE_MEDICAL,BASE_MEDICAL};
   };
   class Tank {
     TransportWeapons[] = {WEST_AT};
@@ -56,10 +57,10 @@ class blu_f {
   class Ship_F {};
 
   class B_Soldier_F {// rifleman
-    uniform[] = {"SAM_Crye_M90_uniform","SAM_Crye_M90_uniform_r"};
-    vest[] = {"SAM_PlateCarrier"};
-    headgear[] = {"SAM_HelmetIA_M90"};
-    backpack[] = {"SAM_AssaultPack_rgr"};
+    uniform[] = {"rhs_uniform_cu_ucp"};
+    vest[] = {"rhsusf_iotv_ucp_Rifleman"};
+    headgear[] = {"rhsusf_ach_helmet_ucp","rhsusf_ach_helmet_ESS_ucp"};
+    backpack[] = {"rhsusf_assault_eagleaiii_ucp"};
     backpackItems[] = {BASE_MEDICAL};
     weapons[] = {WEST_RIFLE};
     magazines[] = {WEST_RIFLE_MAG,WEST_BASE_GRENADES};
@@ -72,12 +73,15 @@ class blu_f {
     magazines[] = {WEST_CARBINE_MAG,WEST_BASE_GRENADES};
   };
   class B_Soldier_TL_F: B_Soldier_F {// FTL
+    vest[] = {"rhsusf_iotv_ucp_Grenadier"};
     weapons[] = {WEST_GLRIFLE};
+    headgear[] = {"rhsusf_ach_helmet_headset_ess_ucp"};
     magazines[] = {WEST_GLRIFLE_MAG,WEST_GLRIFLE_MAG_HE,WEST_GLRIFLE_MAG_SMOKE,WEST_LEADER_GRENADES};
     items[] += {WEST_LEADER_TOOLS};
     linkedItems[] += {WEST_LEADER_LINKED,BINOS};
   };
   class B_Soldier_SL_F: B_Soldier_TL_F {// SL
+    headgear[] = {"rhsusf_ach_helmet_headset_ucp"};
     handguns[] = {WEST_PISTOL};
     magazines[] += {WEST_PISTOL_MAG};
     linkedItems[] = {WEST_BASE_LINKED,WEST_LEADER_LINKED,RANGE_FINDER};
@@ -91,6 +95,7 @@ class blu_f {
     linkedItems[] += {"B_uavterminal"};
   };
   class B_Soldier_AR_F: B_Soldier_F {// AR
+    vest[] = {"rhsusf_iotv_ucp_SAW"};
     weapons[] = {WEST_AR};
     magazines[] = {WEST_AR_MAG,WEST_PISTOL_MAG,WEST_BASE_GRENADES};
     handguns[] = {WEST_PISTOL};
@@ -105,7 +110,7 @@ class blu_f {
     launchers[] = {WEST_AT};
   };
   class B_medic_F: Fic_Soldier_Carbine {// Medic
-    backpack[] = {"SAM_Kitbag_rgr"};
+    vest[] = {"rhsusf_iotv_ucp_medic"};
     magazines[] = {WEST_CARBINE_MAG,MEDIC_GRENADES};
     backpackItems[] = {MEDIC_MEDICAL};
   };
@@ -113,29 +118,27 @@ class blu_f {
     linkedItems[] += {RANGE_FINDER};
   };
   class B_support_MG_F: B_Soldier_AR_F {// MMG
-    backpack[] = {"SAM_Kitbag_rgr"};
     weapons[] = {WEST_MMG};
     magazines[] = {WEST_MMG_MAG,WEST_PISTOL_MAG,WEST_BASE_GRENADES};
     attachments[] = {};
   };
   class B_Soldier_A_F: Fic_Spotter {// MMG Spotter/Ammo Bearer
-    backpack[] = {"SAM_Kitbag_rgr"};
-    backpackItems[] += {WEST_MMG_MAG};
+    backpackItems[] += {WEST_MMG_MAG2};
   };
   class B_soldier_AT_F: Fic_Soldier_Carbine {// MAT Gunner
-    LAUNCHER_GEAR("SAM_GRG_Backpack_Heat", WEST_MAT_MAG)
+    LAUNCHER_GEAR("B_Carryall_mcamo", WEST_MAT_MAG)
     launchers[] = {WEST_MAT};
     secondaryAttachments[] = {WEST_MAT_OPTIC};
   };
   class B_Soldier_AAT_F: Fic_Spotter {// MAT Spotter/Ammo Bearer
-    LAUNCHER_GEAR("SAM_GRG_Backpack_Heat", WEST_MAT_MAG)
+    LAUNCHER_GEAR("B_Carryall_mcamo", WEST_MAT_MAG)
   };
   class B_soldier_AA_F: Fic_Soldier_Carbine {// SAM Gunner
-    LAUNCHER_GEAR("B_Carryall_oli", WEST_SAM_MAG)
+    LAUNCHER_GEAR("B_Carryall_mcamo", WEST_SAM_MAG)
     launchers[] = {WEST_SAM};
   };
   class B_Soldier_AAA_F: Fic_Spotter {// SAM Spotter/Ammo Bearer
-    LAUNCHER_GEAR("B_Carryall_oli", WEST_SAM_MAG2)
+    LAUNCHER_GEAR("B_Carryall_mcamo", WEST_SAM_MAG2)
   };
   class B_support_Mort_F: Fic_Soldier_Carbine {// Mortar Gunner
     MORTAR_GEAR("B_Mortar_01_weapon_F")
@@ -144,8 +147,8 @@ class blu_f {
     MORTAR_GEAR("B_Mortar_01_support_F")
   };
   class B_spotter_F: Fic_Spotter {// Spotter
-    uniform[] = {"SAM_Crye_M90_ghillie"};
-    headgear[] = {"SAM_Watchcap_grn"};
+    vest[] = {"rhsusf_iotv_ucp"};
+    headgear[] = {"rhs_Booniehat_ucp"};
     weapons[] = {WEST_SPOTTER};
     magazines[] = {WEST_SPOTTER_MAG,WEST_BASE_GRENADES};
     items[] += {RADIO_MR,"ACE_ATragMX","ACE_Kestrel4500","ACE_RangeCard"};
@@ -171,17 +174,22 @@ class blu_f {
   };
   class B_helicrew_F: B_Helipilot_F {}; // Pilot
   class B_crew_F: Fic_Soldier_Carbine {// Crew
-    backpack[] = {"SAM_Kitbag_rgr_eng"};
+    vest[] = {"rhsusf_iotv_ucp"};
+    headgear[] = {"rhsusf_cvc_green_ess"};
+    backpack[] = {"B_Carryall_mcamo"};
     magazines[] = {WEST_CARBINE_MAG,CREW_GRENADES};
     backpackItems[] = {KEY_WEST,RADIO_LR};
     linkedItems[] = {WEST_BASE_LINKED,WEST_LEADER_LINKED,BINOS};
+    items[] += {BASE_MEDICAL};
   };
   class B_soldier_repair_F: B_crew_F {// Repair Specialist
-    backpack[] = {"SAM_Kitbag_rgr_eng"};
+    backpack[] = {"B_Carryall_mcamo"};
     backpackItems[] = {"Toolkit",RADIO_MR,KEY_WEST};
+    vest[] = {"rhsusf_iotv_ucp_repair"};
     linkedItems[] = {WEST_BASE_LINKED,WEST_LEADER_LINKED};
   };
   class Fic_eng: B_soldier_repair_F {
+    headgear[] = {"rhsusf_ach_helmet_ucp","rhsusf_ach_helmet_ESS_ucp"};
     items[] += {BASE_ENG};
     backpackItems[] = {};
   };
