+++ f/spect/fn_ForceExit.sqf
@@ -11,7 +11,7 @@ if (isNil "_respawn" || !_respawn) then {
 };
 
 {
-  _var = _x getVariable ["f_cam_fired_eventid",nil];
+  _var = _x getVariable ["f_cam_fired_eventid", nil];
   if(!isNil "_var") then {
     _x removeEventHandler ["fired",_var];
   };
